+++ HISTORY.md
@@ -1,9 +1,14 @@
-## 0.4.0 (in progress)
+## 0.5.0 (in progress)
+
+
+## 0.4.0
 
 - Fix for handling unicode, thanks to @chapmanb and @sowmyaiyer
 - Adds support for adding BAM tags to aligned fastqtransformed files. Thanks to @chapmanb.
 - Adds support for UMI-only fastqtransformation.
 - Adds support for paired-end target sequences.
+- Adds support for detecting sample barcodes via the SB tag in the regex.
+- Adds support for sample-based demultiplexing with error correction.
 
 ## 0.3.0